+++ src/main/java/com/builtbroken/atomic/content/machines/processing/extractor/TileEntityChemExtractor.java
@@ -2,8 +2,6 @@ package com.builtbroken.atomic.content.machines.processing.extractor;
 
 import com.builtbroken.atomic.AtomicScience;
 import com.builtbroken.atomic.client.EffectRefs;
-import com.builtbroken.atomic.content.ASBlocks;
-import com.builtbroken.atomic.content.ASItems;
 import com.builtbroken.atomic.content.machines.processing.ProcessorRecipeHandler;
 import com.builtbroken.atomic.content.machines.processing.TileEntityProcessingMachine;
 import com.builtbroken.atomic.content.machines.processing.extractor.gui.ContainerExtractor;
@@ -16,7 +14,6 @@ import cpw.mods.fml.common.network.ByteBufUtils;
 import io.netty.buffer.ByteBuf;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.inventory.ISidedInventory;
-import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraftforge.common.util.ForgeDirection;
@@ -274,8 +271,7 @@ public class TileEntityChemExtractor extends TileEntityProcessingMachine impleme
     {
         if (slot == SLOT_ITEM_INPUT)
         {
-            return stack.getItem() == Item.getItemFromBlock(ASBlocks.blockUraniumOre)
-                    || stack.getItem() == ASItems.itemProcessingWaste;
+            return getRecipeList().isComponent(this, stack);
         }
         return false;
     }