+++ src/main/java/com/builtbroken/atomic/content/machines/processing/centrifuge/gui/ContainerChemCentrifuge.java
@@ -1,12 +1,11 @@
 package com.builtbroken.atomic.content.machines.processing.centrifuge.gui;
 
+import com.builtbroken.atomic.content.items.wrench.WrenchColor;
 import com.builtbroken.atomic.content.machines.processing.centrifuge.TileEntityChemCentrifuge;
 import com.builtbroken.atomic.lib.gui.ContainerBase;
 import com.builtbroken.atomic.lib.gui.slot.MachineSlot;
+import com.builtbroken.atomic.lib.gui.slot.SlotOutput;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.inventory.SlotFurnace;
-
-import java.awt.*;
 
 /**
  * @see <a href="https://github.com/BuiltBrokenModding/VoltzEngine/blob/development/license.md">License</a> for what you can and can't do with the code.
@@ -17,13 +16,13 @@ public class ContainerChemCentrifuge extends ContainerBase<TileEntityChemCentrif
     public ContainerChemCentrifuge(EntityPlayer player, TileEntityChemCentrifuge tile)
     {
         super(player, tile);
-        addSlotToContainer(new MachineSlot(tile, TileEntityChemCentrifuge.SLOT_FLUID_INPUT, 25, 52).setColor(Color.blue));
-        addSlotToContainer(new MachineSlot(tile, TileEntityChemCentrifuge.SLOT_FLUID_OUTPUT, 136, 52).setColor(Color.GREEN));
+        addSlotToContainer(new MachineSlot(tile, TileEntityChemCentrifuge.SLOT_FLUID_INPUT, 25, 52).setColor(WrenchColor.BLUE.getColor()));
+        addSlotToContainer(new MachineSlot(tile, TileEntityChemCentrifuge.SLOT_FLUID_OUTPUT, 136, 52).setColor(WrenchColor.GREEN.getColor()));
 
-        addSlotToContainer(new SlotFurnace(player, tile, TileEntityChemCentrifuge.SLOT_ITEM_OUTPUT, 100, 30));
+        addSlotToContainer(new SlotOutput(tile, TileEntityChemCentrifuge.SLOT_ITEM_OUTPUT, 100, 30).setColor(WrenchColor.ORANGE.getColor()));
 
         int x = 50;
-        addSlotToContainer(new MachineSlot(tile, TileEntityChemCentrifuge.SLOT_BATTERY, x, 52));
+        addSlotToContainer(new MachineSlot(tile, TileEntityChemCentrifuge.SLOT_BATTERY, x, 52).setColor(WrenchColor.PURPLE.getColor()));
         addPlayerInventory(player);
     }
 }